Only remove contents of zwift data directory
See https://github.com/netbrain/zwift/issues/375.

+++ src/update_zwift.sh
@@ -161,8 +161,7 @@ cleanup() {
     rm -- "/tmp/webview2-setup.exe" || true
     rm -rf -- "${WINE_USER_HOME}/Downloads/Zwift" || true
     rm -rf -- "/home/user/.cache/wine*" || true
-    # remove Zwift documents because it causes permission errors with podman
-    rm -rf -- "${ZWIFT_DATA_DIR}" || true
+    rm -rf -- "${ZWIFT_DATA_DIR:?}/*" || true
 }
 
 trap cleanup EXIT
